reorg cluster component
